refactor(service): MemberCategoryTemplateApplicationService 삭제
