+++ backend/src/main/java/codezap/category/controller/CategoryController.java
@@ -19,8 +19,8 @@ import codezap.category.dto.request.CreateCategoryRequest;
 import codezap.category.dto.request.UpdateCategoryRequest;
 import codezap.category.dto.response.CreateCategoryResponse;
 import codezap.category.dto.response.FindAllCategoriesResponse;
+import codezap.category.service.CategoryTemplateService;
 import codezap.category.service.facade.MemberCategoryApplicationService;
-import codezap.category.service.facade.MemberCategoryTemplateApplicationService;
 import codezap.global.validation.ValidationSequence;
 import codezap.member.domain.Member;
 import lombok.RequiredArgsConstructor;
@@ -31,7 +31,7 @@ import lombok.RequiredArgsConstructor;
 public class CategoryController implements SpringDocCategoryController {
 
     private final MemberCategoryApplicationService memberCategoryApplicationService;
-    private final MemberCategoryTemplateApplicationService memberCategoryTemplateApplicationService;
+    private final CategoryTemplateService categoryTemplateService;
 
     @PostMapping
     public ResponseEntity<CreateCategoryResponse> createCategory(
@@ -39,14 +39,11 @@ public class CategoryController implements SpringDocCategoryController {
             @Validated(ValidationSequence.class) @RequestBody CreateCategoryRequest createCategoryRequest
     ) {
         CreateCategoryResponse createdCategory = memberCategoryApplicationService.create(member, createCategoryRequest);
-        return ResponseEntity.created(URI.create("/categories/" + createdCategory.id()))
-                .body(createdCategory);
+        return ResponseEntity.created(URI.create("/categories/" + createdCategory.id())).body(createdCategory);
     }
 
     @GetMapping
-    public ResponseEntity<FindAllCategoriesResponse> getCategories(
-            @RequestParam Long memberId
-    ) {
+    public ResponseEntity<FindAllCategoriesResponse> getCategories(@RequestParam Long memberId) {
         return ResponseEntity.ok(memberCategoryApplicationService.findAllByMember(memberId));
     }
 
@@ -62,8 +59,7 @@ public class CategoryController implements SpringDocCategoryController {
 
     @DeleteMapping("/{id}")
     public ResponseEntity<Void> deleteCategory(@AuthenticationPrinciple Member member, @PathVariable Long id) {
-        memberCategoryTemplateApplicationService.deleteById(member, id);
-        return ResponseEntity.noContent()
-                .build();
+        categoryTemplateService.deleteById(member, id);
+        return ResponseEntity.noContent().build();
     }
 }
