+++ backend/src/test/java/codezap/global/MockMvcTest.java
@@ -9,7 +9,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
-import org.springframework.context.annotation.Import;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.setup.MockMvcBuilders;
@@ -19,12 +18,11 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 
 import codezap.auth.manager.CredentialManager;
 import codezap.auth.provider.CredentialProvider;
+import codezap.category.service.CategoryTemplateService;
 import codezap.category.service.facade.MemberCategoryApplicationService;
-import codezap.category.service.facade.MemberCategoryTemplateApplicationService;
 import codezap.global.cors.CorsProperties;
 import codezap.member.fixture.MemberFixture;
 import codezap.member.service.MemberService;
-import codezap.template.service.TemplateService;
 
 @WebMvcTest(SpringExtension.class)
 @EnableConfigurationProperties(CorsProperties.class)
@@ -42,13 +40,10 @@ public abstract class MockMvcTest {
     protected CredentialManager credentialManager;
 
     @MockBean
-    protected TemplateService templateService;
-
-    @MockBean
     protected MemberCategoryApplicationService memberCategoryApplicationService;
 
     @MockBean
-    protected MemberCategoryTemplateApplicationService memberCategoryTemplateApplicationService;
+    protected CategoryTemplateService categoryTemplateService;
 
     @MockBean
     protected MemberService memberService;
